L.346
Add a camera bar button to initiate the ImagePicker Controller, add an IBAction to it "cameraButtonTapped"
Add an imageView to display the image the user has chosen, add an IBOutlet to it "imageView"

Add user description inside the info.plist file for using the camera and the photo library.

Add imagePicker Controller that will be displayed when the user taps on the camera bar button, return the image the user has chosen from their photo library.
Cast the userImage as a CIImage and send it to a function called detect.

detect() will send the image to the model and print the results received.